Make it possible to pass run param via args
A proper default.properties file is needed to run the integration test.
For that, one can simply create the file but for the CI is impossible to
do. However the CI can totally inject the properties via the command
line. This method makes it possible and will create a default.properties
file or update an existing one with the properties injected.