Replace ip with adbSerial in GenymotionVirtualDevice
We need an adb serial to talk to cloud devices because all cloud devices
appear on localhost as far as adb is concerned, but they use different
ports.

Parsing the adb serial field from `gmtool admin details` failed because the
code was splitting on ':' without limiting the number of splits. So I
modified all the splits for key-value lines to limit their splits to 2.