Create a helloworld project for gradle
We used to link a commit to binocle with a submodule to demonstrate gradle.
This make things painful as we have "fork" branch in binocle (as we don't want
the gradle plugin in binocle master).
To make things simpler, we demonstrate the gradle plugin with a very simple
helloworld project