Fix plugin tests
Some tests were failing because they required a mips64el-linux-android ABI.
It is no longer supported. Fixed by updating the Android gradle plugin
to 3.2.0 by default. Before it was set to '+', meaning latest available
version, but google repo was not set, so an old version was fetched.

More generally, it's a bad idea to use a '+', so I choose a fixed version.
It required Gradle version 4.6 at least.